Move tests and fixtures from tests/src to tests/tests
