--- tests/src/lib.rs
@@ -1,5 +1,0 @@
-#[cfg(test)]
-mod fixtures;
-
-#[cfg(test)]
-mod tests;
